+++ simplivity/resources/policies.py
@@ -98,6 +98,11 @@ class Policy(object):
         self._connection = connection
         self._client = resource_client
 
+    def __refresh(self):
+        """Updates the policy data."""
+        resource_uri = "{}/{}".format(URL, self.data["id"])
+        self.data = self._client.do_get(resource_uri)['policy']
+
     def get_vms(self):
         """Retrieves the virtual machines using this policy.
 
@@ -120,3 +125,50 @@ class Policy(object):
         resource_uri = "{}/{}".format(URL, self.data["id"])
         self._client.do_delete(resource_uri, timeout, None)
         self.data = None
+
+    def create_rules(self, rules, replace_all_rules=False, timeout=-1):
+        """Creates one or more new rules or replaces existing rules with new rules for a policy
+
+        Args:
+            rules: Array of rules.(below are the parameter that can be passed in the array)
+                [Mandatory Parameters]
+                frequency: The number of minutes between backups
+                retention: The number of minutes to keep backups
+                [Optional Parameters]
+                application_consistent: Set false for crash-consistent backups
+                                        Set true for application-consistent backups (for example, VSS or snapshot backups)
+                                        Default: false
+                consistency_type: Set to DEFAULT for a snapshot backup
+                                  Set to VSS for a Microsoft Volume Shadow Copy Service backup
+                                  Set to NONE for crash-consistent backups
+                                  Default: NONE
+                destination_id: The unique identifier (UID) of the omnistack_cluster to store the backup
+                                Default: local omnistack_cluster
+                days: The days of the week (for example, Mon,Fri), or month (for example, 1,15) to take backups or
+                      "last" to specify the last day of each month
+                      Default: All (that is, every day)
+                start_time: The time to start the backups, for example, 14:30
+                            This time is local to the time zone of the Hypervisor Management System (HMS)
+                            Default: 00:00
+                end_time: The time to stop backing up, for example, 14:30
+                          This time is local to the time zone of the Hypervisor Management System (HMS)
+                          Default: 00:00
+                external_store_name: The name of the external_store
+
+            replace_all_rules: If set to True, replaces the existing rules with new rules
+                               If set to False, adds the new rules to existing set of rules
+            timeout: Time out for the request in seconds.
+
+
+        Returns:
+            self: Returns the policy object.
+        """
+        resource_uri = "{}/{}/rules".format(URL, self.data["id"])
+        if isinstance(rules, dict):
+            rules = [rules]
+
+        flags = {'replace_all_rules': replace_all_rules}
+        self._client.do_post(resource_uri, rules, timeout, None, flags)[0]
+        self.__refresh()
+
+        return self